+++ plugin/src/main/bash/data.bin
@@ -1,1 +1,1 @@
-BUILD_NUM="1257"
+BUILD_NUM="1260"
