+++ plugin/src/main/java/com/craftmend/openaudiomc/spigot/OpenAudioMcSpigot.java
@@ -6,7 +6,7 @@ import com.craftmend.openaudiomc.api.interfaces.AudioApi;
 import com.craftmend.openaudiomc.generic.environment.MagicValue;
 import com.craftmend.openaudiomc.generic.logging.OpenAudioLogger;
 import com.craftmend.openaudiomc.generic.logging.platform.GenericLogAdapter;
-import com.craftmend.openaudiomc.generic.proxy.SpigotUserHooks;
+import com.craftmend.openaudiomc.spigot.modules.users.SpigotUserHooks;
 import com.craftmend.openaudiomc.generic.proxy.interfaces.UserHooks;
 import com.craftmend.openaudiomc.generic.state.StateService;
 import com.craftmend.openaudiomc.generic.storage.interfaces.Configuration;
