+++ plugin/src/main/java/com/craftmend/openaudiomc/spigot/modules/commands/command/SpigotAudioCommand.java
@@ -19,7 +19,7 @@ import com.craftmend.openaudiomc.generic.state.StateService;
 import com.craftmend.openaudiomc.generic.state.interfaces.State;
 import com.craftmend.openaudiomc.generic.state.states.WorkerState;
 import com.craftmend.openaudiomc.generic.user.User;
-import com.craftmend.openaudiomc.generic.user.adapters.SpigotUserAdapter;
+import com.craftmend.openaudiomc.spigot.modules.users.adapters.SpigotUserAdapter;
 import com.craftmend.openaudiomc.spigot.modules.events.SpigotAudioCommandEvent;
 import com.craftmend.openaudiomc.spigot.modules.players.objects.SpigotPlayerSelector;
