+++ plugin/src/main/java/com/craftmend/openaudiomc/spigot/modules/users/SpigotUserHooks.java
@@ -1,4 +1,4 @@
-package com.craftmend.openaudiomc.generic.proxy;
+package com.craftmend.openaudiomc.spigot.modules.users;
 
 import com.craftmend.openaudiomc.OpenAudioMc;
 import com.craftmend.openaudiomc.generic.networking.interfaces.NetworkingService;
@@ -8,8 +8,8 @@ import com.craftmend.openaudiomc.generic.proxy.messages.StandardPacket;
 import com.craftmend.openaudiomc.generic.proxy.models.ProxyNode;
 import com.craftmend.openaudiomc.generic.user.User;
 import com.craftmend.openaudiomc.generic.user.adapters.CommandSenderUserAdapter;
-import com.craftmend.openaudiomc.generic.user.adapters.SpigotUserAdapter;
 import com.craftmend.openaudiomc.spigot.modules.proxy.service.ProxyNetworkingService;
+import com.craftmend.openaudiomc.spigot.modules.users.adapters.SpigotUserAdapter;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
