+++ plugin/src/main/resources/openaudiomc-build.properties
@@ -1,3 +1,3 @@
-BUILD_VERSION="1257"
-BUILD_COMMIT="34c85282ae6421662ea2a1509f74c9f16e82d3e2"
+BUILD_VERSION="1260"
+BUILD_COMMIT="ca516c91c267109bcd1a0f3a4617e8368035aaf1"
 BUILD_AUTHOR="Mats"
